requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
clickhouse-connect==0.6.18
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
idna==3.4
lz4==4.3.2
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
pytz==2023.3.post1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0
zstandard==0.22.0
